# docs/swagger.yaml
basePath: /
definitions:
  buildinfo.Info:
    properties:
      build_time:
        type: string
      commit:
        type: string
      started_at:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
  health.Dependency:
    properties:
      latency:
        type: string
      status:
        type: string
    type: object
  health.Status:
    properties:
      dependencies:
        additionalProperties:
          $ref: '#/definitions/health.Dependency'
        type: object
      status:
        type: string
    type: object
  httptransport.BatchCheckAuthorizationRequest:
    properties:
      checks:
        items:
          $ref: '#/definitions/httptransport.CheckAuthorizationRequest'
        maxItems: 100
        minItems: 1
        type: array
    required:
    - checks
    type: object
  httptransport.CheckAuthorizationRequest:
    properties:
      action:
        type: string
      attributes:
        additionalProperties:
          type: string
        type: object
      resource_id:
        type: string
      resource_type:
        type: string
      subject_id:
        type: string
      subject_type:
        type: string
      tenant_id:
        type: string
    required:
    - action
    - resource_type
    - subject_id
    - subject_type
    - tenant_id
    type: object
  httptransport.CreateBindingRequest:
    properties:
      organization_unit_id:
        type: string
      role_id:
        type: string
      subject_id:
        type: string
      subject_type:
        type: string
      tenant_id:
        type: string
    required:
    - role_id
    - subject_id
    - subject_type
    - tenant_id
    type: object
  httptransport.CreatePermissionRequest:
    properties:
      action:
        type: string
      code:
        type: string
      condition_expression:
        type: string
      name:
        type: string
      resource_type:
        type: string
      tenant_id:
        type: string
    required:
    - action
    - code
    - name
    - resource_type
    - tenant_id
    type: object
  httptransport.CreateRoleRequest:
    properties:
      code:
        type: string
      data_scope:
        type: string
      description:
        type: string
      name:
        type: string
      tenant_id:
        type: string
    required:
    - code
    - data_scope
    - name
    - tenant_id
    type: object
  httptransport.GrantRolePermissionRequest:
    properties:
      permission_id:
        type: string
      role_id:
        type: string
      tenant_id:
        type: string
    required:
    - permission_id
    - role_id
    - tenant_id
    type: object
  httptransport.ListBindingsRequest:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      subject_id:
        type: string
      subject_type:
        type: string
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
  httptransport.ListPermissionsRequest:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
  httptransport.ListRolePermissionsRequest:
    properties:
      role_id:
        type: string
    required:
    - role_id
    type: object
  httptransport.ListRolesRequest:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
  httptransport.MeResponseBody:
    properties:
      subject:
        type: string
    type: object
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
  httptransport.RevokeBindingRequest:
    properties:
      binding_id:
        type: string
      version:
        type: integer
    required:
    - binding_id
    - version
    type: object
  httptransport.RevokeRolePermissionRequest:
    properties:
      role_permission_id:
        type: string
      version:
        type: integer
    required:
    - role_permission_id
    - version
    type: object
  httptransport.UpdateRoleRequest:
    properties:
      data_scope:
        type: string
      description:
        type: string
      name:
        type: string
      role_id:
        type: string
      status:
        type: string
      version:
        type: integer
    required:
    - data_scope
    - name
    - role_id
    - status
    - version
    type: object
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
  /api/v1/authorization/batch-check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization checks
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.BatchCheckAuthorizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Preview up to 100 authorization decisions
      tags:
      - authorization-decisions
  /api/v1/authorization/bindings/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role binding
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CreateBindingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Bind a role to a membership, group, or service account
      tags:
      - authorization-bindings
  /api/v1/authorization/bindings/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant, optional subject, and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListBindingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: List tenant or subject role bindings
      tags:
      - authorization-bindings
  /api/v1/authorization/bindings/revoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: Binding and current version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.RevokeBindingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Revoke a binding with optimistic locking
      tags:
      - authorization-bindings
  /api/v1/authorization/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CheckAuthorizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Preview an authorization decision
      tags:
      - authorization-decisions
  /api/v1/authorization/permissions/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Permission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Create a tenant permission
      tags:
      - authorization-permissions
  /api/v1/authorization/permissions/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: List tenant permissions
      tags:
      - authorization-permissions
  /api/v1/authorization/role-permissions/grant:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role permission
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GrantRolePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Grant a permission to a role
      tags:
      - authorization-roles
  /api/v1/authorization/role-permissions/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: List permissions assigned to a role
      tags:
      - authorization-roles
  /api/v1/authorization/role-permissions/revoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role permission and current version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.RevokeRolePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Revoke a role permission with optimistic locking
      tags:
      - authorization-roles
  /api/v1/authorization/roles/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Create a tenant role
      tags:
      - authorization-roles
  /api/v1/authorization/roles/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: List tenant roles
      tags:
      - authorization-roles
  /api/v1/authorization/roles/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role and current version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Update a role with optimistic locking
      tags:
      - authorization-roles
  /api/v1/me:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.MeResponseBody'
              type: object
        "401":
          description: 'Code 20001: unauthorized'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Return the authenticated subject
      tags:
      - authentication
  /api/v1/version:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/buildinfo.Info'
              type: object
      summary: Return build and runtime version information
      tags:
      - operations
  /live:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
      summary: Check process liveness
      tags:
      - operations
  /ready:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
        "503":
          description: 'Code 50003: dependency unavailable'
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/health.Status'
              type: object
      summary: Check database and Redis readiness
      tags:
      - operations
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"